Move `parseBoolean` to utils
